A few tweeks on Movies(API)

+++ IEC.API/Persistence/Repositories/MovieArtistRepository.cs
@@ -1,3 +1,4 @@
+using System.Collections.Generic;
 using System.Linq;
 using IEC.API.Core.Domain;
 using IEC.API.Core.Repositories;
@@ -8,9 +9,11 @@ namespace IEC.API.Persistence.Repositories
     {
         public MovieArtistRepository(DataContext context) :base(context) {}
 
-        public void DeleteMovieArtists(int movieId)
+        public void DeleteMovieArtists(int movieId, List<int> artistIds, List<int> roleIds)
         {
-            Context.RemoveRange(Context.MovieArtists.Where(m => movieId == m.MovieId));
+            // Context.RemoveRange(Context.MovieArtists.Where(m => movieId == m.MovieId));
+            for(var i = 0; i < artistIds.Count; i++) 
+                Context.RemoveRange(Context.MovieArtists.Where(m => movieId == m.MovieId && artistIds[i] == m.ArtistId && roleIds[i] == m.RoleId));
         }
     }
 }
